LayoutTests/imported/w3c/web-platform-tests/svg/animations/animateMotion-keyPoints-001-expected.txt


FAIL 'calcMode' paced and 'keyPoints' on <animateMotion> with 'path' assert_equals: y position expected 0 but got -50
